requirements.txt
PyQt6>=6.4.0
tushare>=1.2.89
mysql-connector-python>=8.0.0
pandas>=1.5.0
numpy>=1.23.0
apscheduler>=3.9.1
matplotlib>=3.6.0
python-dotenv>=1.0.0